# feat: implement standardized error handling and unified structured logging