--- a/pkg/mcp/types.go
+++ b/pkg/mcp/types.go
@@ -33,3 +33,10 @@ type Relation struct {
 	CreatedAt   time.Time `json:"createdAt"`
 	UpdatedAt   time.Time `json:"updatedAt"`
 }
+
+// Error represents a JSON-RPC error response compatible with MCP transport layer
+type Error struct {
+	Code    int         `json:"code"`
+	Message string      `json:"message"`
+	Data    interface{} `json:"data,omitempty"`
+}