… ERROR_HANDLING.md
… LOGGING.md
… PHASE2_IMPLEMENTATION_SUMMARY.md
… PHASE4_IMPLEMENTATION.md
… README.md
… cmd/mcp-server/integration_test.go
… cmd/mcp-server/main_integration_test.go
… cmd/mcp-server/main_test.go
… cmd/mcp-server/server_edge_cases_test.go
… cmd/mcp-server/stdio_test.go
… config.logging.yaml
… example_audit.log
… examples/config.logging.yaml
… examples/logging_example.go
… go.sum
… internal/admin/server.go
… internal/knowledge/manager.go
… internal/knowledge/manager_edge_cases_test.go
… internal/knowledge/manager_json_test.go
… internal/storage/error_integration_test.go
… internal/storage/factory_edge_cases_test.go
… internal/storage/factory_test.go
… internal/storage/memory.go
… internal/storage/memory_edge_cases_test.go
… internal/storage/memory_test.go
… internal/storage/sqlite.go
… internal/storage/sqlite_test.go
… internal/transport/error_mapper.go
… internal/transport/error_mapper_test.go
… internal/transport/race_test.go